feat(cart) - 장바구니 관련 기능 연결

+++ src/main/java/com/lionkit/mogumarket/cart/dto/reaponse/CartItemResponse.java
@@ -3,7 +3,7 @@ package com.lionkit.mogumarket.cart.dto.reaponse;
 public record CartItemResponse(
         Long productId,
         String productName,
-        Integer unitPrice,
+        long unitPrice,
         Integer quantity,
-        Integer lineTotal // = unitPrice * quantity
+        long lineTotal // = unitPrice * quantity
 ) {}
